バトル終了のイベントコマンドを追加
