+++ src/rmmz/eventCommand/commands/battle/battleProcessing/index.ts
@@ -1,19 +1,4 @@
+export * from "./make";
+export * from "./types";
+
 export { BATTLE_DIRECT, BATTLE_ENCOUNT, BATTLE_VARIABLE } from "./constants";
-
-export type {
-  Command_BattleProcessing,
-  Command_BattleProcessing_Variable,
-  ParamArray_BattleProcessing,
-  ParamArray_BattleProcessing_Direct,
-  ParamArray_BattleProcessing_Encount,
-  ParamArray_BattleProcessing_Variable,
-  ParamObject_BattleProcessing_Direct,
-  ParamObject_BattleProcessing_Variable,
-} from "./types";
-
-export {
-  isCommandBattleProcessingVariable,
-  makeCommandBattleProcessingDirect,
-  makeCommandBattleProcessingEncount,
-  makeCommandBattleProcessingVariable,
-} from "./make";
