+++ src/rmmz/eventCommand/commands/battle/battleProcessing/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./battleIf";
+export * from "./startBattle";
