+++ src/rmmz/eventCommand/commands/union.ts
@@ -38,12 +38,13 @@ import type {
 } from "./audio/play/types";
 import type { Command_StopSE } from "./audio/stop";
 import type { Command_ShowBalloonIcon } from "./balloon/types";
-import type { Command_AbortBattle } from "./battle/abort/types";
-import type { Command_ChangeBattleBackground } from "./battle/background/types";
 import type {
   Command_BattleProcessing,
+  Command_BattleProcessing_BlockEnd,
   Command_BattleProcessing_If,
-} from "./battle/battleProcessing/types";
+} from "./battle";
+import type { Command_AbortBattle } from "./battle/abort/types";
+import type { Command_ChangeBattleBackground } from "./battle/background/types";
 import type { Command_ForceAction } from "./battle/forceAction/types/union";
 import type { Command_ShowBattleAnimation } from "./battle/showAnimation/types";
 import type { Command_ControlTimer } from "./controlTimer/types";
@@ -188,6 +189,7 @@ export type NormalizedEventCommand =
   | Command_ChangeBattleBackground
   | Command_BattleProcessing
   | Command_BattleProcessing_If
+  | Command_BattleProcessing_BlockEnd
   | Command_ControlTimer
   | Command_ChangeEnemyHP
   | Command_ChangeEnemyMP
